requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
mysql-connector-python==8.2.0
# face-recognition==1.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
numpy>=1.21.0,<2.0.0
Pillow==10.1.0
setuptools>=65.0.0
cmake
dlib-bin


